+++ examples/express/handler.js
@@ -1,23 +1,22 @@
-import {App} from '../../common/universal/components/app';
+import {Provider} from 'react-redux';
 import React from 'react';
 import {RoutingContext} from 'react-router';
-import {getState} from '../../common/universal/state';
-import {renderToStaticStyle} from '../../common/style';
+import {createServerStore} from '../store';
+import {renderToStaticStyle} from '../style';
 import {renderToString} from 'react-dom/server';
 
-export function handler ({props, reply, request}) {
+export function handler ({props, request, response}) {
   const style = renderToStaticStyle({userAgent: request.headers['user-agent']});
 
-  getState(props, state => {
+  createServerStore(props, store => {
     const app = renderToString(
-      <App state={state}>
+      <Provider store={store}>
         <RoutingContext {...props}/>
-      </App>
+      </Provider>
     );
+    const state = JSON.stringify(store.getState());
 
-    state = JSON.stringify(state);
-
-    reply(`<!doctype html>
+    response.send(`<!doctype html>
       <html>
         <head>
           <meta charSet="utf-8"/>
@@ -26,7 +25,7 @@ export function handler ({props, reply, request}) {
         <body>
           <div id="app">${app}</div>
           <script id="state" type="application/json">${state}</script>
-          <script src="/example-universal.js"></script>
+          <script src="/client.js"></script>
         </body>
       </html>
     `);