+++ examples/hapi/index.js
@@ -1,7 +1,10 @@
 import {Server} from 'hapi';
-import {plugin as api} from './api';
-import {plugin as ruto} from './ruto';
-import {plugin as webpack} from './webpack';
+import {assets, compiler} from '../webpack';
+import {handler} from './handler';
+import {register as registerAPI} from './api';
+import registerWebpack from 'hapi-webpack-plugin';
+import {register as registerRuto} from 'ruto';
+import {routes} from '../routes';
 
 const {env} = process;
 const host = env.npm_config_host || '0.0.0.0';
@@ -10,10 +13,14 @@ const server = new Server();
 
 server.connection({host, port});
 
-server.register([api, ruto, webpack], error => {
+server.register([
+  {register: registerAPI},
+  {options: {handler, routes}, register: registerRuto},
+  {options: {assets, compiler}, register: registerWebpack}
+], error => {
   if (error) { throw error; }
 
   server.start(() => {
-    console.log(`Rūto hapi: ${server.info.uri}`);
+    console.log(`rūto hapi: ${server.info.uri}`);
   });
 });