+++ src/ruto/middleware.js
@@ -1,21 +1,13 @@
 import {createLocation} from 'history';
 import {match} from 'react-router';
 
-export function middleware (options) {
-  const {cache = false, getHandler, getRoutes} = options;
-  let {handler, routes} = options;
-
-  if (!handler && !getHandler) { throw new Error('Missing handler.'); }
-  if (!routes && !getRoutes) { throw new Error('Missing routes.'); }
-  if (cache && getHandler) { handler = getHandler(); }
-  if (cache && getRoutes) { routes = getRoutes(); }
+export function middleware ({handler, routes}) {
+  if (!handler) { throw new Error('Missing handler.'); }
+  if (!routes) { throw new Error('Missing routes.'); }
 
   return function use (request, response, next) {
-    const {query, url} = request;
+    const {url} = request;
     const location = createLocation(url);
-
-    if (!cache && getHandler) { handler = getHandler(query); }
-    if (!cache && getRoutes) { routes = getRoutes(query); }
 
     match({location, routes}, (error, redirect, props) => {
       if (error) { return next(error); }